20240913 company

+++ src/main/java/company/demo_last_last202409121553/login/repository/LoginUserRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface LoginUserRepository extends JpaRepository<LoginUser, Long> {
+public interface LoginUserRepository extends JpaRepository<LoginUser, Integer> {
 
     Optional<LoginUser> findByEmail(String email);
 }
